Refactored CardDeck.js
Refactored Card and Paginate Buttons into separate files. Card is in Components/CardDeck and Paginate Buttons is in UI.
Moved CardDeck itself into its own folder.
